move useHideSplashScreen from /helpers to /hooks directory
