+++ src/hooks/index.js
@@ -29,6 +29,7 @@ export { default as useEmailRainbow } from './useEmailRainbow';
 export { default as useExpandedStateNavigation } from './useExpandedStateNavigation';
 export { default as useGas } from './useGas';
 export { default as useHeight } from './useHeight';
+export { default as useHideSplashScreen } from './useHideSplashScreen';
 export { default as useImageMetadata } from './useImageMetadata';
 export { default as useInitializeAccountData } from './useInitializeAccountData';
 export { default as useInitializeWallet } from './useInitializeWallet';
